requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
replicate
asyncio 